refactor: Separate logic to model
